AWS_AIChatbot/requirements.txt
langchain-astradb 
langchain_core
python-dotenv
streamlit
langchain_community
datasets
langchain-openai
pypdf
flask
groq
langchain-groq
